Updated date format for yMd skeleton, because for electronic data Poland is using the ISO 8601 format since 2002.
The traditional date format shouldn't be used on web interfaces actually... Using the traditional one will break up with the defaults of .NET framework as example, which expects the ISO8601 format for dates if you set the culture format for Poland. Yet this format is nowhere to be found in here? I've chosen the yMd skeleton, because this is also the default for jQuery Validation.